fix(mod_arith): widen bound type to uint64_t for 64-bit storage support
Goldilocks field (p = 2⁶⁴ - 2³² + 1) requires i64 storage where
getMaxBound() can exceed 2³²-1. Using unsigned truncated the bound,
causing incorrect lazy reduction analysis.

+++ prime_ir/Dialect/ModArith/Conversions/ModArithToArith/Reducer/MontReducer.h
@@ -50,7 +50,7 @@ public:
   // For bound <= 1 returns the input as-is; for bound == 2 uses a single
   // conditional subtraction; for bound > 2 uses binary conditional subtraction
   // in ceil(log₂(bound)) steps.
-  Value getCanonicalFromExtended(Value input, unsigned bound = 2);
+  Value getCanonicalFromExtended(Value input, uint64_t bound = 2);
 
   // Gets the canonical form from an input value in [0, 2n) and an overflow
   // flag.
